added unit tests for pullOpenapiSpecs both in GeneratorTest and SpecExtractorHelperTest
