+++ src/test/java/com/gs/crdtools/GeneratorTest.java
@@ -40,4 +40,23 @@ public class GeneratorTest {
         assertTrue(result.get(cronTabSpec).get().contains("class CronTabSpec"));
         assertTrue(result.get(cronTab).get().contains("class CronTab"));
     }
+
+    @Test
+    void testGenerateCodeContainsAdditionalProperties() throws IOException  {
+        // Additionally to SpecExtractorHelperTest.testPullOpenapiSpecsAddsInfo()
+        // this test checks if the additional properties are added to the generated code
+
+        var runFiles = Runfiles.create();
+        var input = Path.of(runFiles.rlocation("__main__/src/test/resources/minimal-crd.yaml"));
+
+        var result = Generator.generate(Generator.parseCrds(List.of(input)));
+
+        var base = Path.of("src/main/java/com/gs/crdtools");
+        var cronTab = base.resolve("CronTab.java");
+
+
+        assertTrue(result.get(cronTab).get().contains("@JsonProperty(\"metadata\")"));
+        assertTrue(result.get(cronTab).get().contains("@JsonProperty(\"kind\")"));
+        assertTrue(result.get(cronTab).get().contains("@JsonProperty(\"apiVersion\")"));
+    }
 }
